+++ CHANGES.rst
@@ -1,7 +1,7 @@
 Changelog
 =========
 
-6.1.3 (unreleased)
+6.1.4 (unreleased)
 ------------------
 
 Minor changes:
@@ -19,6 +19,14 @@ New features:
 Bug fixes:
 
 - ...
+
+6.1.3 (2025-03-19)
+------------------
+
+Bug fixes:
+
+- Fix to permit TZID forward references to VTIMEZONEs
+- Stabelize timezone id lookup, see `Issue 780 <https://github.com/collective/icalendar/issues/780>`_.
 
 6.1.2 (2025-03-19)
 ------------------